r/kubernetes_cluster_node_pool: moving the Node Pool ID to the parse package
